"Load more" link hidden when datasource defines "maxItems" property
When a datasource sets a positive "maxItems" property in its
description, the template hides the "Load more" link if the returned
feed contains more items than the value specified in the "maxItems".

This is particularly useful for feeds that only return one item
such as the "static content" or the "Youtube video" datasources.